Created the frontend for the application
Made two pages the Register.jsx and the Login.jsx. Styled them and connected them with the backend implementation. When a user logs in and has the role of Admin they will be redirected to the AdminPanel.jsx the other are going to the WelcomePage.jsx.

Created some custom components for a form and resolved routing and centralized api calls in one big service: ApiServices.js

Fixed some error display bugs