Optimized the AdminPanel.jsx. Created the ChooseRole.jsx alongside with ResidentRequest.jsx and the CompanyRequest.jsx
Separated the AdminPanel.jsx into smaller components to be easier to read and to separate concerns. The created components only absorb props and only returns jsx with css. With that improvement the AdminPanel.jsx stays as a parent component with only logic.

Created the role selection part but for now without any logic.